split container client into multiple pkg
