+++ pkg/chaosdaemon/server_test.go
@@ -14,8 +14,13 @@
 package chaosdaemon
 
 import (
+	"context"
+	"github.com/chaos-mesh/chaos-mesh/pkg/chaosdaemon/crclients/test"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
+	"github.com/prometheus/client_golang/prometheus"
+	"golang.org/x/sync/errgroup"
+	"net"
 
 	"github.com/chaos-mesh/chaos-mesh/pkg/mock"
 )
@@ -23,7 +28,7 @@ import (
 var _ = Describe("netem server", func() {
 	Context("newDaemonServer", func() {
 		It("should work", func() {
-			defer mock.With("MockContainerdClient", &MockClient{})()
+			defer mock.With("MockContainerdClient", &test.MockClient{})()
 			_, err := newDaemonServer(containerRuntimeContainerd)
 			Expect(err).To(BeNil())
 		})
@@ -36,17 +41,77 @@ var _ = Describe("netem server", func() {
 
 	Context("newGRPCServer", func() {
 		It("should work", func() {
-			defer mock.With("MockContainerdClient", &MockClient{})()
+			defer mock.With("MockContainerdClient", &test.MockClient{})()
 			_, err := newGRPCServer(containerRuntimeContainerd, &MockRegisterer{})
 			Expect(err).To(BeNil())
 		})
 
 		It("should panic", func() {
 			Ω(func() {
-				defer mock.With("MockContainerdClient", &MockClient{})()
+				defer mock.With("MockContainerdClient", &test.MockClient{})()
 				defer mock.With("PanicOnMustRegister", "mock panic")()
 				newGRPCServer(containerRuntimeContainerd, &MockRegisterer{})
 			}).Should(Panic())
 		})
 	})
 })
+
+type MockRegisterer struct {
+	RegisterGatherer
+}
+
+func (*MockRegisterer) MustRegister(...prometheus.Collector) {
+	if err := mock.On("PanicOnMustRegister"); err != nil {
+		panic(err)
+	}
+}
+
+
+// RegisterGatherer combine prometheus.Registerer and prometheus.Gatherer
+type RegisterGatherer interface {
+	prometheus.Registerer
+	prometheus.Gatherer
+}
+
+// StartServer starts chaos-daemon.
+func StartServer(conf *Config, reg RegisterGatherer) error {
+	g := &errgroup.Group{}
+
+	httpBindAddr := conf.HttpAddr()
+	httpServer := newHTTPServerBuilder().Addr(httpBindAddr).Metrics(reg).Profiling(conf.Profiling).Build()
+
+	grpcBindAddr := conf.GrpcAddr()
+	grpcListener, err := net.Listen("tcp", grpcBindAddr)
+	if err != nil {
+		log.Error(err, "failed to listen grpc address", "grpcBindAddr", grpcBindAddr)
+		return err
+	}
+
+	grpcServer, err := newGRPCServer(conf.Runtime, reg)
+	if err != nil {
+		log.Error(err, "failed to create grpc server")
+		return err
+	}
+
+	g.Go(func() error {
+		log.Info("Starting http endpoint", "address", httpBindAddr)
+		if err := httpServer.ListenAndServe(); err != nil {
+			log.Error(err, "failed to start http endpoint")
+			httpServer.Shutdown(context.Background())
+			return err
+		}
+		return nil
+	})
+
+	g.Go(func() error {
+		log.Info("Starting grpc endpoint", "address", grpcBindAddr, "runtime", conf.Runtime)
+		if err := grpcServer.Serve(grpcListener); err != nil {
+			log.Error(err, "failed to start grpc endpoint")
+			grpcServer.Stop()
+			return err
+		}
+		return nil
+	})
+
+	return g.Wait()
+}
