+++ pkg/chaosdaemon/util.go
@@ -16,15 +16,14 @@ package chaosdaemon
 import (
 	"context"
 	"fmt"
+	"github.com/chaos-mesh/chaos-mesh/pkg/chaosdaemon/crclients/docker"
 	"io/ioutil"
 	"net/http"
 	"os"
 	"strconv"
 	"sync"
-	"syscall"
 
-	"github.com/containerd/containerd"
-	"github.com/docker/docker/api/types"
+	"github.com/chaos-mesh/chaos-mesh/pkg/chaosdaemon/crclients/containerd"
 	dockerclient "github.com/docker/docker/client"
 
 	"github.com/chaos-mesh/chaos-mesh/pkg/bpm"
@@ -35,12 +34,9 @@ const (
 	containerRuntimeDocker     = "docker"
 	containerRuntimeContainerd = "containerd"
 
-	defaultDockerSocket  = "unix:///var/run/docker.sock"
-	dockerProtocolPrefix = "docker://"
-
 	// TODO(yeya24): make socket and ns configurable
+	defaultDockerSocket  = "unix:///var/run/docker.sock"
 	defaultContainerdSocket  = "/run/containerd/containerd.sock"
-	containerdProtocolPrefix = "containerd://"
 	containerdDefaultNS      = "k8s.io"
 )
 
@@ -51,110 +47,19 @@ type ContainerRuntimeInfoClient interface {
 	FormatContainerID(ctx context.Context, containerID string) (string, error)
 }
 
-// DockerClientInterface represents the DockerClient, it's used to simply unit test
-type DockerClientInterface interface {
-	ContainerInspect(ctx context.Context, containerID string) (types.ContainerJSON, error)
-	ContainerKill(ctx context.Context, containerID, signal string) error
-}
-
-// DockerClient can get information from docker
-type DockerClient struct {
-	client DockerClientInterface
-}
-
-// FormatContainerID strips protocol prefix from the container ID
-func (c DockerClient) FormatContainerID(ctx context.Context, containerID string) (string, error) {
-	if len(containerID) < len(dockerProtocolPrefix) {
-		return "", fmt.Errorf("container id %s is not a docker container id", containerID)
-	}
-	if containerID[0:len(dockerProtocolPrefix)] != dockerProtocolPrefix {
-		return "", fmt.Errorf("expected %s but got %s", dockerProtocolPrefix, containerID[0:len(dockerProtocolPrefix)])
-	}
-	return containerID[len(dockerProtocolPrefix):], nil
-}
-
-// GetPidFromContainerID fetches PID according to container id
-func (c DockerClient) GetPidFromContainerID(ctx context.Context, containerID string) (uint32, error) {
-	id, err := c.FormatContainerID(ctx, containerID)
-	if err != nil {
-		return 0, err
-	}
-	container, err := c.client.ContainerInspect(ctx, id)
-	if err != nil {
-		return 0, err
-	}
-
-	if container.State.Pid == 0 {
-		return 0, fmt.Errorf("container is not running, status: %s", container.State.Status)
-	}
-
-	return uint32(container.State.Pid), nil
-}
-
-// ContainerdClientInterface represents the ContainerClient, it's used to simply unit test
-type ContainerdClientInterface interface {
-	LoadContainer(ctx context.Context, id string) (containerd.Container, error)
-}
-
-// ContainerdClient can get information from containerd
-type ContainerdClient struct {
-	client ContainerdClientInterface
-}
-
-// FormatContainerID strips protocol prefix from the container ID
-func (c ContainerdClient) FormatContainerID(ctx context.Context, containerID string) (string, error) {
-	if len(containerID) < len(containerdProtocolPrefix) {
-		return "", fmt.Errorf("container id %s is not a containerd container id", containerID)
-	}
-	if containerID[0:len(containerdProtocolPrefix)] != containerdProtocolPrefix {
-		return "", fmt.Errorf("expected %s but got %s", containerdProtocolPrefix, containerID[0:len(containerdProtocolPrefix)])
-	}
-	return containerID[len(containerdProtocolPrefix):], nil
-}
-
-// GetPidFromContainerID fetches PID according to container id
-func (c ContainerdClient) GetPidFromContainerID(ctx context.Context, containerID string) (uint32, error) {
-	id, err := c.FormatContainerID(ctx, containerID)
-	if err != nil {
-		return 0, err
-	}
-	container, err := c.client.LoadContainer(ctx, id)
-	if err != nil {
-		return 0, err
-	}
-	task, err := container.Task(ctx, nil)
-	if err != nil {
-		return 0, err
-	}
-	return task.Pid(), nil
-}
 
 // newDockerclient returns a dockerclient.NewClient with mock points
-func newDockerClient(host string, version string, client *http.Client, httpHeaders map[string]string) (DockerClientInterface, error) {
+func newDockerClient(host string, version string, client *http.Client, httpHeaders map[string]string) (docker.DockerClientInterface, error) {
 	// Mock point to return error or mock client in unit test
 	if err := mock.On("NewDockerClientError"); err != nil {
 		return nil, err.(error)
 	}
 	if client := mock.On("MockDockerClient"); client != nil {
-		return client.(DockerClientInterface), nil
+		return client.(docker.DockerClientInterface), nil
 	}
 
 	// The real logic
 	return dockerclient.NewClient(host, version, client, httpHeaders)
-}
-
-// newContainerdClient returns a containerd.New with mock points
-func newContainerdClient(address string, opts ...containerd.ClientOpt) (ContainerdClientInterface, error) {
-	// Mock point to return error in unit test
-	if err := mock.On("NewContainerdClientError"); err != nil {
-		return nil, err.(error)
-	}
-	if client := mock.On("MockContainerdClient"); client != nil {
-		return client.(ContainerdClientInterface), nil
-	}
-
-	// The real logic
-	return containerd.New(address, opts...)
 }
 
 // CreateContainerRuntimeInfoClient creates a container runtime information client.
@@ -162,21 +67,20 @@ func CreateContainerRuntimeInfoClient(containerRuntime string) (ContainerRuntime
 	// TODO: support more container runtime
 
 	var cli ContainerRuntimeInfoClient
+	var err error
 	switch containerRuntime {
 	case containerRuntimeDocker:
-		client, err := newDockerClient(defaultDockerSocket, "", nil, nil)
+		cli, err = docker.New(defaultDockerSocket, "", nil, nil)
 		if err != nil {
 			return nil, err
 		}
-		cli = DockerClient{client}
 
 	case containerRuntimeContainerd:
 		// TODO(yeya24): add more options?
-		client, err := newContainerdClient(defaultContainerdSocket, containerd.WithDefaultNamespace(containerdDefaultNS))
+		cli, err = containerd.New(defaultContainerdSocket, containerd.WithDefaultNamespace(containerdDefaultNS))
 		if err != nil {
 			return nil, err
 		}
-		cli = ContainerdClient{client}
 
 	default:
 		return nil, fmt.Errorf("only docker and containerd is supported, but got %s", containerRuntime)
@@ -184,43 +88,6 @@ func CreateContainerRuntimeInfoClient(containerRuntime string) (ContainerRuntime
 
 	return cli, nil
 }
-
-// ContainerKillByContainerID kills container according to container id
-func (c DockerClient) ContainerKillByContainerID(ctx context.Context, containerID string) error {
-	if len(containerID) < len(dockerProtocolPrefix) {
-		return fmt.Errorf("container id %s is not a docker container id", containerID)
-	}
-	if containerID[0:len(dockerProtocolPrefix)] != dockerProtocolPrefix {
-		return fmt.Errorf("expected %s but got %s", dockerProtocolPrefix, containerID[0:len(dockerProtocolPrefix)])
-	}
-	err := c.client.ContainerKill(ctx, containerID[len(dockerProtocolPrefix):], "SIGKILL")
-
-	return err
-}
-
-// ContainerKillByContainerID kills container according to container id
-func (c ContainerdClient) ContainerKillByContainerID(ctx context.Context, containerID string) error {
-	if len(containerID) < len(containerdProtocolPrefix) {
-		return fmt.Errorf("container id %s is not a containerd container id", containerID)
-	}
-	if containerID[0:len(containerdProtocolPrefix)] != containerdProtocolPrefix {
-		return fmt.Errorf("expected %s but got %s", containerdProtocolPrefix, containerID[0:len(containerdProtocolPrefix)])
-	}
-	containerID = containerID[len(containerdProtocolPrefix):]
-	container, err := c.client.LoadContainer(ctx, containerID)
-	if err != nil {
-		return err
-	}
-	task, err := container.Task(ctx, nil)
-	if err != nil {
-		return err
-	}
-
-	err = task.Kill(ctx, syscall.SIGKILL)
-
-	return err
-}
-
 // ReadCommName returns the command name of process
 func ReadCommName(pid int) (string, error) {
 	f, err := os.Open(fmt.Sprintf("%s/%d/comm", bpm.DefaultProcPrefix, pid))
